Consolidate Lambda functions
