--- cmd/lambda/clearpositions/dockerfile
@@ -1,23 +1,0 @@
-FROM public.ecr.aws/lambda/python:3.11-arm64
-
-RUN yum clean all
-
-RUN yum install -y gcc python3-devel
-
-COPY cmd/lambda/clearpositions/requirements.txt ${LAMBDA_TASK_ROOT}/requirements.txt
-
-RUN pip3 install --requirement requirements.txt --target "${LAMBDA_TASK_ROOT}"
-
-COPY pkg/trade/trade.py ${LAMBDA_TASK_ROOT}/pkg/trade/trade.py
-
-COPY cmd/lambda/clearpositions/main.py ${LAMBDA_TASK_ROOT}/main.py
-
-ENV DARQUBE_API_KEY=""
-
-ENV ALPACA_API_KEY=""
-
-ENV ALPACA_API_SECRET=""
-
-ENV IS_PAPER=""
-
-CMD [ "main.handler" ]
