--- cmd/lambda/clearpositions/requirements.txt
@@ -1,3 +1,0 @@
-alpaca-py
-requests
-numpy
