--- cmd/lambda/createpositions/requirements.txt
@@ -1,12 +1,0 @@
-alpaca-py
-pandas
-requests
-boto3
-sagemaker
-scikit-learn
-numpy
-tensorflow
-keras
-wandb
-beautifulsoup4
-lxml # required for beautifulsoup4
